+++ views/admin/motion-type/_sections.php
@@ -182,17 +182,12 @@ $sName = 'sections[' . $sectionId . ']';
         </div>
 
         <?php
-        /**
-         * @param int $i
-         * @param string $sectionName
-         * @return string
-         */
-        $dataRowFormatter = function (TabularDataType $row, $i, $sectionName) {
+        $dataRowFormatter = function (TabularDataType $row, int $i, string $sectionName): string {
             $str = '<li class="no' . $i . '">';
             $str .= '<span class="drag-data-handle">&#9776;</span>';
             $str .= '<input type="text" name="' . $sectionName . '[tabular][' . $row->rowId . '][title]"';
-            $str .= ' placeholder="Angabe" value="' . Html::encode($row->title) . '" class="form-control">';
-            $str .= '<select name="' . $sectionName . '[tabular][' . $row->rowId . '][type]" class="form-control">';
+            $str .= ' placeholder="' . Yii::t('admin', 'motion_section_tab_p') . '" value="' . Html::encode($row->title) . '" class="form-control">';
+            $str .= '<select name="' . $sectionName . '[tabular][' . $row->rowId . '][type]" class="form-control tabularTypeSelect">';
             foreach (TabularDataType::getDataTypes() as $dataId => $dataName) {
                 $str .= '<option value="' . $dataId . '"';
                 if ($row->type == $dataId) {
@@ -201,7 +196,19 @@ $sName = 'sections[' . $sectionId . ']';
                 $str .= '>' . Html::encode($dataName) . '</option>';
             }
             $str .= '</select>';
-            $str .= '<a href="#" class="delRow glyphicon glyphicon-remove-circle"></a>';
+            $str .= '<button type="button" class="delRow btn btn-link" title="' . Yii::t('admin', 'motion_section_rowdel') . '">';
+            $str .= '<span class="glyphicon glyphicon-remove-circle" aria-hidden="true"></span>';
+            $str .= '<span class="sr-only">' . Yii::t('admin', 'motion_section_rowdel') . '</span>';
+            $str .= '</button>';
+            $str .= '<div class="selectOptions hidden">';
+            $str .= '<div class="description">' . Yii::t('admin', 'motion_section_options') . ':</div>';
+            $str .= '<div class="selectize-wrapper">';
+            $str .= '<select class="tags" name="' . $sectionName . '[tabular][' . $row->rowId . '][options][]" multiple="multiple">';
+            foreach ($row->options as $option) {
+                $str .= '<option value="' . Html::encode($option) . '" selected>' . Html::encode($option) . '</option>';
+            }
+            $str .= '</select></div>';
+            $str .= '</div>';
             $str .= '</li>';
             return $str;
         };
@@ -229,10 +236,10 @@ $sName = 'sections[' . $sectionId . ']';
             ]);
             $template = $dataRowFormatter($newRow, 0, $sName);
             ?>
-            <a href="#" class="addRow" data-template="<?= Html::encode($template) ?>">
+            <button type="button" class="btn btn-link addRow" data-template="<?= Html::encode($template) ?>">
                 <span class="glyphicon glyphicon-plus-sign" aria-hidden="true"></span>
                 <?= Yii::t('admin', 'motion_section_add_line') ?>
-            </a>
+            </button>
         </fieldset>
 
     </div>